+++ mod-source-record-manager-server/src/test/java/org/folio/rest/impl/changeManager/ChangeManagerAPITest.java
@@ -30,6 +30,7 @@ import org.folio.rest.jaxrs.model.InitialRecord;
 import org.folio.rest.jaxrs.model.JobExecution;
 import org.folio.rest.jaxrs.model.JobProfile;
 import org.folio.rest.jaxrs.model.JobProfileInfo;
+import org.folio.rest.jaxrs.model.JobProfileInfo.DataType;
 import org.folio.rest.jaxrs.model.ProfileSnapshotWrapper;
 import org.folio.rest.jaxrs.model.Progress;
 import org.folio.rest.jaxrs.model.RawRecordsDto;
@@ -194,7 +195,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     requestDto.setSourceType(InitJobExecutionsRqDto.SourceType.ONLINE);
     requestDto.setJobProfileInfo(new JobProfileInfo()
       .withId(DEFAULT_JOB_PROFILE_ID)
-      .withDataType(JobProfileInfo.DataType.MARC)
+      .withDataType(JobProfileInfo.DataType.MARC_BIB)
       .withName("Test Profile"));
 
     InitJobExecutionsRsDto response = RestAssured.given()
@@ -958,7 +959,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(DEFAULT_JOB_PROFILE_ID)
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(JobProfileInfo.DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -1017,7 +1018,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(DEFAULT_JOB_PROFILE_ID)
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(JobProfileInfo.DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -1076,7 +1077,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(DEFAULT_JOB_PROFILE_ID)
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(JobProfileInfo.DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -1180,7 +1181,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(DEFAULT_JOB_PROFILE_ID)
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(JobProfileInfo.DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -1235,7 +1236,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(DEFAULT_JOB_PROFILE_ID)
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(JobProfileInfo.DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -1283,7 +1284,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(DEFAULT_JOB_PROFILE_ID)
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(JobProfileInfo.DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -1333,7 +1334,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(DEFAULT_JOB_PROFILE_ID)
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(JobProfileInfo.DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -1477,7 +1478,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(DEFAULT_JOB_PROFILE_ID)
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(JobProfileInfo.DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -1549,7 +1550,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(DEFAULT_JOB_PROFILE_ID)
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(JobProfileInfo.DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -1581,7 +1582,22 @@ public class ChangeManagerAPITest extends AbstractRestTest {
   }
 
   @Test
-  public void shouldFillInRecordOrderIfAtLeastOneRecordHasNoOrder() throws InterruptedException, IOException {
+  public void shouldFillInRecordOrderIfAtLeastOneMarcBibRecordHasNoOrder() throws InterruptedException, IOException {
+    fillInRecordOrderIfAtLeastOneRecordHasNoOrder(DataType.MARC_BIB);
+  }
+
+  @Test
+  public void shouldFillInRecordOrderIfAtLeastOneMarcAuthorityRecordHasNoOrder() throws InterruptedException, IOException {
+    fillInRecordOrderIfAtLeastOneRecordHasNoOrder(DataType.MARC_AUTHORITY);
+  }
+
+  @Test
+  public void shouldFillInRecordOrderIfAtLeastOneMarcAuthorityMarcHoldingRecordHasNoOrder() throws InterruptedException, IOException {
+    fillInRecordOrderIfAtLeastOneRecordHasNoOrder(DataType.MARC_HOLDING);
+  }
+
+  private void fillInRecordOrderIfAtLeastOneRecordHasNoOrder(DataType dataType)
+    throws InterruptedException, IOException {
     RawRecordsDto rawRecordsDto = new RawRecordsDto()
       .withId(UUID.randomUUID().toString())
       .withRecordsMetadata(new RecordsMetadata()
@@ -1603,7 +1619,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(DEFAULT_JOB_PROFILE_ID)
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(dataType))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -1625,7 +1641,8 @@ public class ChangeManagerAPITest extends AbstractRestTest {
     Event obtainedEvent = Json.decodeValue(observedValues.get(0), Event.class);
     assertEquals(DI_RAW_RECORDS_CHUNK_PARSED.value(), obtainedEvent.getEventType());
 
-    RecordCollection processedRecords = Json.decodeValue(ZIPArchiver.unzip(obtainedEvent.getEventPayload()), RecordCollection.class);
+    RecordCollection processedRecords = Json
+      .decodeValue(ZIPArchiver.unzip(obtainedEvent.getEventPayload()), RecordCollection.class);
     assertEquals(3, processedRecords.getRecords().size());
 
     assertEquals(4, processedRecords.getRecords().get(0).getOrder().intValue());
@@ -1653,7 +1670,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(DEFAULT_JOB_PROFILE_ID)
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(JobProfileInfo.DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -1723,7 +1740,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(DEFAULT_JOB_PROFILE_ID)
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(JobProfileInfo.DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -1778,7 +1795,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(DEFAULT_JOB_PROFILE_ID)
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(JobProfileInfo.DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -1822,7 +1839,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(JOB_PROFILE_ID)
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(JobProfileInfo.DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -1869,7 +1886,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(DEFAULT_JOB_PROFILE_ID)
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(JobProfileInfo.DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -1919,7 +1936,7 @@ public class ChangeManagerAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(DEFAULT_JOB_PROFILE_ID)
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()