+++ mod-source-record-manager-server/src/test/java/org/folio/rest/impl/changeManager/ChangeManagerParsedRecordsAPITest.java
@@ -69,6 +69,7 @@ public class ChangeManagerParsedRecordsAPITest extends AbstractRestTest {
   }
 
   @Test
+  @Ignore("Need to remove ignore and uncomment recordType when MODSOURCE-279 will be merge in SRS module")
   public void shouldReturnParsedRecordDtoIfSourceRecordExists(TestContext testContext) {
     Async async = testContext.async();
 
@@ -77,7 +78,7 @@ public class ChangeManagerParsedRecordsAPITest extends AbstractRestTest {
       .withRecordId(UUID.randomUUID().toString())
       .withParsedRecord(new ParsedRecord().withId(UUID.randomUUID().toString())
         .withContent("{\"leader\":\"01240cas a2200397   4500\",\"fields\":[]}"))
-      .withRecordType(SourceRecord.RecordType.MARC)
+//      .withRecordType(SourceRecord.RecordType.MARC_BIB)
       .withExternalIdsHolder(new ExternalIdsHolder().withInstanceId(instanceId));
 
     WireMock.stubFor(get(new UrlPathPattern(new RegexPattern(SOURCE_RECORDS_URL + ".*"), true))
@@ -162,7 +163,7 @@ public class ChangeManagerParsedRecordsAPITest extends AbstractRestTest {
       .withId(UUID.randomUUID().toString())
       .withParsedRecord(new ParsedRecord().withId(UUID.randomUUID().toString())
         .withContent("{\"leader\":\"01240cas a2200397   4500\",\"fields\":[]}"))
-      .withRecordType(ParsedRecordDto.RecordType.MARC)
+      .withRecordType(ParsedRecordDto.RecordType.MARC_BIB)
       .withExternalIdsHolder(new ExternalIdsHolder().withInstanceId(UUID.randomUUID().toString()));
 
     RestAssured.given()
@@ -180,5 +181,27 @@ public class ChangeManagerParsedRecordsAPITest extends AbstractRestTest {
       .build());
     async.complete();
   }
+    @Test
+    public void shouldReturnErrorOnPutIfFailedToSendEvent(TestContext testContext) {
+      Async async = testContext.async();
 
+      ParsedRecordDto parsedRecordDto = new ParsedRecordDto()
+        .withId(UUID.randomUUID().toString())
+        .withParsedRecord(new ParsedRecord().withId(UUID.randomUUID().toString())
+          .withContent("{\"leader\":\"01240cas a2200397   4500\",\"fields\":[]}"))
+        .withRecordType(ParsedRecordDto.RecordType.MARC_BIB)
+        .withExternalIdsHolder(new ExternalIdsHolder().withInstanceId(UUID.randomUUID().toString()));
+
+      WireMock.stubFor(post(PUBSUB_PUBLISH_URL)
+        .willReturn(WireMock.serverError()));
+
+      RestAssured.given()
+        .spec(spec)
+        .body(parsedRecordDto)
+        .when()
+        .put(PARSED_RECORDS_URL + "/" + parsedRecordDto.getId())
+        .then()
+        .statusCode(HttpStatus.SC_INTERNAL_SERVER_ERROR);
+      async.complete();
+    }
 }