+++ mod-source-record-manager-server/src/test/java/org/folio/rest/impl/metadataProvider/MetadataProviderJobExecutionAPITest.java
@@ -15,6 +15,7 @@ import org.folio.rest.jaxrs.model.JobExecution;
 import org.folio.rest.jaxrs.model.JobExecutionCollection;
 import org.folio.rest.jaxrs.model.JobExecutionLogDto;
 import org.folio.rest.jaxrs.model.JobProfileInfo;
+import org.folio.rest.jaxrs.model.JobProfileInfo.DataType;
 import org.folio.rest.jaxrs.model.JournalRecord.ActionType;
 import org.folio.rest.jaxrs.model.JournalRecord.EntityType;
 import org.folio.rest.jaxrs.model.JournalRecordCollection;
@@ -253,7 +254,7 @@ public class MetadataProviderJobExecutionAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(UUID.randomUUID().toString())
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(JobProfileInfo.DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -312,7 +313,7 @@ public class MetadataProviderJobExecutionAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(UUID.randomUUID().toString())
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(JobProfileInfo.DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -362,7 +363,7 @@ public class MetadataProviderJobExecutionAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(UUID.randomUUID().toString())
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(JobProfileInfo.DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -428,7 +429,7 @@ public class MetadataProviderJobExecutionAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(UUID.randomUUID().toString())
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(JobProfileInfo.DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -511,7 +512,7 @@ public class MetadataProviderJobExecutionAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(UUID.randomUUID().toString())
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(JobProfileInfo.DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()
@@ -617,7 +618,7 @@ public class MetadataProviderJobExecutionAPITest extends AbstractRestTest {
       .body(new JobProfileInfo()
         .withName("MARC records")
         .withId(UUID.randomUUID().toString())
-        .withDataType(JobProfileInfo.DataType.MARC))
+        .withDataType(DataType.MARC_BIB))
       .when()
       .put(JOB_EXECUTION_PATH + jobExec.getId() + JOB_PROFILE_PATH)
       .then()